Changed "ms" to "μs" for reporting query time
The abbreviation "ms" is milliseconds while the abbreviation "μs" is microseconds.  There are 1,000 microseconds in a millesecond.  The value of the variable it is referencing is also called `total_query_execution_time_in_micros`.  I did a cursory check of the logic behind this variable to verify it _is_ reporting in microseconds:

`query_execution_time_in_micros = int((end_time - start_time) * 1000 * 1000)`

`end_time - start_time` is in seconds, multiply by 1,000 is in milleseconds and multiplying again by 1,000 is in microseconds. 👍

While I was in these files, I also fixed some casing of words/acronyms that I saw.